plug(gb_col): plug the collector int the codebase parsing phase.

+++ heredoc/heredoc.c
@@ -10,7 +10,6 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/signals.h"
 #include "../minishell.h"
 
 static void	run_heredoc_child(t_heredoc *hd, t_minishell *minishell)
@@ -19,7 +18,8 @@ static void	run_heredoc_child(t_heredoc *hd, t_minishell *minishell)
 	while (1)
 	{
 		hd->line = readline("> ");
-		if (!hd->line || strcmp(hd->line, hd->del) == 0)
+		collector_register(hd->line, SCOPE_CHILD);
+		if (!hd->line || (hd->line && ft_strcmp(hd->line, hd->del) == 0))
 			break ;
 		if (hd->quote == NQS)
 		{
@@ -27,13 +27,12 @@ static void	run_heredoc_child(t_heredoc *hd, t_minishell *minishell)
 		}
 		write(hd->tmp_fd, hd->line, ft_strlen(hd->line));
 		write(hd->tmp_fd, "\n", 1);
-		free(hd->line);
 	}
-	free(hd->line);
 	close(hd->tmp_fd);
+	close(hd->fd);
 	free(hd->filename);
 	free(hd->del);
-	exit(EXIT_SUCCESS);
+	cleanup_exit(0);
 }
 
 static int	run_heredoc_parent(t_heredoc *hd)
@@ -49,37 +48,38 @@ static int	run_heredoc_parent(t_heredoc *hd)
 			ft_putstr_fd("\n", 1);
 			g_received_signal = SIGNAL_SIGINT;
 		}
-		return (1);
+		return (130);
 	}
 	else if (WIFEXITED(hd->status) && WEXITSTATUS(hd->status) != 0)
-		return (perror("open tmp for reading"), 1);
+		return (1);
 	return (0);
 }
 
 int	heredoc(t_heredoc *hd, t_minishell *minishell)
 {
 	hd->tmp_fd = open(hd->filename, O_CREAT | O_WRONLY, 0600);
+	if (hd->tmp_fd == -1)
+		return (1);
 	hd->fd = open(hd->filename, O_RDONLY, 0600);
+	return (1);
 	unlink(hd->filename);
 	hd->pid = fork();
 	if (hd->pid < 0)
-	{
-		perror("fork");
-		return (-1);
-	}
+		return (perror("fork"), 1);
 	if (hd->pid == 0)
 		run_heredoc_child(hd, minishell);
 	else
 	{
 		signal(SIGINT, SIG_IGN);
-		if (run_heredoc_parent(hd) != 0)
+		minishell->status = run_heredoc_parent(hd);
+		if (minishell->status == 130)
 		{
 			setup_shell_signals();
-            ft_putstr_fd(HEREDOC_ERROR, 2);
-            minishell->status = 130;
+			ft_putstr_fd(HEREDOC_ERROR, 2);
+			minishell->status = 130;
 			return (1);
 		}
 		setup_shell_signals();
 	}
-	return (0);
+	return (minishell->status);
 }
